refactor: move region state into separate module
