+++ src/mito2/src/request.rs
@@ -55,8 +55,8 @@ use crate::manifest::action::{RegionEdit, TruncateKind};
 use crate::memtable::MemtableId;
 use crate::memtable::bulk::part::BulkPart;
 use crate::metrics::COMPACTION_ELAPSED_TOTAL;
-use crate::region::RegionRequestPolicyGuard;
 use crate::region::options::RegionOptions;
+use crate::region::state::RegionRequestPolicyGuard;
 use crate::sst::file::FileMeta;
 use crate::sst::index::IndexBuildType;
 use crate::wal::EntryId;
