+++ src/mito2/src/test_util/scheduler_util.rs
@@ -35,7 +35,8 @@ use crate::config::MitoConfig;
 use crate::error::Result;
 use crate::flush::FlushScheduler;
 use crate::manifest::manager::{RegionManifestManager, RegionManifestOptions};
-use crate::region::{ManifestContext, ManifestContextRef, RegionControlState};
+use crate::region::state::RegionControlState;
+use crate::region::{ManifestContext, ManifestContextRef};
 use crate::request::WorkerRequestWithTime;
 use crate::schedule::scheduler::{Job, LocalScheduler, Scheduler, SchedulerRef};
 use crate::sst::FormatType;
